+++ custom_components/preheat/__init__.py
@@ -36,7 +36,7 @@ async def async_setup_entry(hass: HomeAssistant, entry: PreheatConfigEntry) -> b
 
     await hass.config_entries.async_forward_entry_setups(entry, PLATFORMS)
     entry.async_on_unload(entry.add_update_listener(async_reload_entry))
-
+    
     return True
 
 async def async_unload_entry(hass: HomeAssistant, entry: PreheatConfigEntry) -> bool: